refactor: designComponent 이전에 따른 의존 모듈 import 갱신 Related to: #117
