+++ src/main/java/com/komentum/post/domain/DesignBoard.java
@@ -1,6 +1,6 @@
 package com.komentum.post.domain;
 
-import com.komentum.theme.component.domain.DesignComponent;
+import com.komentum.designcomponent.domain.DesignComponent;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
