+++ src/main/java/com/komentum/post/service/transaction/DesignBoardTransactionService.java
@@ -17,8 +17,8 @@ import com.komentum.post.repository.DesignBoardRepositorySupport;
 import com.komentum.post.service.DesignBoardService;
 import com.komentum.post.service.PostService;
 import com.komentum.post.service.TagService;
-import com.komentum.theme.component.domain.DesignComponent;
-import com.komentum.theme.component.service.DesignComponentService;
+import com.komentum.designcomponent.domain.DesignComponent;
+import com.komentum.designcomponent.service.DesignComponentService;
 import com.komentum.user.domain.User;
 import com.komentum.user.service.UserEntityFinder;
 import jakarta.persistence.EntityNotFoundException;
