+++ src/main/java/com/komentum/seed/seeder/ThemeComponentSeeder.java
@@ -1,10 +1,10 @@
 package com.komentum.seed.seeder;
 
 import com.github.javafaker.Faker;
-import com.komentum.theme.component.domain.ComponentType;
-import com.komentum.theme.component.domain.DesignComponent;
-import com.komentum.theme.component.repository.ComponentTypeRepository;
-import com.komentum.theme.component.repository.DesignComponentRepository;
+import com.komentum.designcomponent.domain.ComponentType;
+import com.komentum.designcomponent.domain.DesignComponent;
+import com.komentum.designcomponent.repository.ComponentTypeRepository;
+import com.komentum.designcomponent.repository.DesignComponentRepository;
 import com.komentum.theme.theme.domain.ThemeComponent;
 import com.komentum.theme.theme.domain.ThemeImage;
 import com.komentum.theme.theme.repository.ThemeComponentRepository;
