+++ src/main/java/com/komentum/theme/theme/domain/ThemeImage.java
@@ -1,7 +1,7 @@
 package com.komentum.theme.theme.domain;
 
-import com.komentum.theme.component.domain.ComponentType;
-import com.komentum.theme.component.domain.DesignComponent;
+import com.komentum.designcomponent.domain.ComponentType;
+import com.komentum.designcomponent.domain.DesignComponent;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
