+++ src/test/java/com/komentum/catalog/controller/ComponentCatalogControllerTest.java
@@ -15,7 +15,7 @@ import com.komentum.test.data.scenario.UserScenarioSupport;
 import com.komentum.test.dto.MockMvcRequestDto;
 import com.komentum.test.dto.TestClientDto;
 import com.komentum.test.dto.TestParams;
-import com.komentum.theme.component.domain.DesignComponent;
+import com.komentum.designcomponent.domain.DesignComponent;
 import com.komentum.user.domain.User;
 import java.util.List;
 import org.junit.jupiter.api.AfterEach;
