+++ src/test/java/com/komentum/test/data/TestDataRemover.java
@@ -8,9 +8,9 @@ import com.komentum.post.repository.PostRepository;
 import com.komentum.post.repository.PreferRepository;
 import com.komentum.post.repository.TagRepository;
 import com.komentum.post.repository.ThemeBoardRepository;
-import com.komentum.theme.component.repository.ColorStyleRepository;
-import com.komentum.theme.component.repository.ComponentTypeRepository;
-import com.komentum.theme.component.repository.DesignComponentRepository;
+import com.komentum.designcomponent.repository.ColorStyleRepository;
+import com.komentum.designcomponent.repository.ComponentTypeRepository;
+import com.komentum.designcomponent.repository.DesignComponentRepository;
 import com.komentum.theme.theme.repository.ThemeComponentRepository;
 import com.komentum.theme.theme.repository.ThemeImageRepository;
 import com.komentum.theme.theme.repository.ThemeStyleRepository;
