+++ src/test/java/com/komentum/theme/android/editor/AndroidThemeMakerTest.java
@@ -10,16 +10,16 @@ import com.komentum.test.config.EnableTestProfile;
 import com.komentum.test.data.ThemeDataGenerator;
 import com.komentum.theme.android.dto.AndroidColorDto;
 import com.komentum.theme.android.dto.AndroidComponentDto;
-import com.komentum.theme.component.domain.ColorStyle;
-import com.komentum.theme.component.domain.ComponentType;
-import com.komentum.theme.component.domain.DesignComponent;
-import com.komentum.theme.component.enums.TypeCode;
+import com.komentum.designcomponent.domain.ColorStyle;
+import com.komentum.designcomponent.domain.ComponentType;
+import com.komentum.designcomponent.domain.DesignComponent;
+import com.komentum.designcomponent.enums.TypeCode;
 import com.komentum.theme.theme.domain.ThemeComponent;
 import com.komentum.theme.theme.domain.ThemeImage;
 import com.komentum.theme.theme.domain.ThemeStyle;
-import com.komentum.theme.utils.ImageUtils;
-import com.komentum.theme.utils.ThemePathManager;
-import com.komentum.theme.utils.XmlEditor;
+import com.komentum.global.utils.ImageUtils;
+import com.komentum.theme.android.utils.ThemePathManager;
+import com.komentum.theme.android.utils.XmlEditor;
 import java.awt.image.BufferedImage;
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
