+++ src/test/java/com/komentum/theme/controller/ThemeManageControllerTest.java
@@ -7,8 +7,8 @@ import com.komentum.test.MockMvcUtils;
 import com.komentum.test.config.EnableTestProfile;
 import com.komentum.test.data.ThemeDataGenerator;
 import com.komentum.test.data.UserDataGenerator;
-import com.komentum.theme.component.dto.CreateThemeRequest;
 import com.komentum.theme.theme.domain.ThemeComponent;
+import com.komentum.theme.theme.dto.CreateThemeRequest;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
